rag-service/requirements.txt
# RAG microservice: FastAPI + LlamaIndex
# Python 3.10+

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LlamaIndex RAG (compatible con Python 3.12)
llama-index
llama-index-core
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-readers-file

# PDF support (SimpleDirectoryReader)
pypdf>=4.0.0
